refactor: move auth business logic to features
